refactor(chat): チャット表示状態の管理をChatHostに集約
- チャットの表示状態とモデル選択状態を ChatHost に集約
- ChatContextProvider と用途別 hook を追加し、dialog と chat UI が必要な状態だけ参照するよう整理
- ChatDialog を新設してダイアログ表示の責務を分離
- OpenChat からダイアログ実装を除き、サイドバーの起動ボタンに責務を限定
